backend/requirements.txt
flask
pathlib
python-dotenv
PyJWT
pillow
Flask-SQLAlchemy
Flask-Migrate
flask-bcrypt
pymysql
cryptography
